Added test files for all extract files
